+++ src/WebJobs.Script.Grpc/Server/Startup.cs
@@ -30,12 +30,16 @@ namespace Microsoft.Azure.WebJobs.Script.Grpc
                 app.UseDeveloperExceptionPage();
             }
 
+            // This must occur before 'UseRouting'. This ensures extension endpoints are registered before the
+            // endpoints are collected by the routing middleware.
+            app.UseMiddleware<ExtensionsCompositeEndpointDataSource.EnsureInitializedMiddleware>();
             app.UseRouting();
 
             app.UseEndpoints(endpoints =>
             {
                 endpoints.MapGrpcService<FunctionRpc.FunctionRpcBase>();
-                endpoints.DataSources.Add(endpoints.ServiceProvider.GetRequiredService<ExtensionsCompositeEndpointDataSource>());
+                endpoints.DataSources.Add(
+                    endpoints.ServiceProvider.GetRequiredService<ExtensionsCompositeEndpointDataSource>());
             });
         }
     }